+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Detect proper USB HID interface to use for CMSIS-DAP v1 probes. Without this, CMSIS-DAP probes with multiple HID interfaces, e.g. MCUlink, were not working properly on MacOS (#722).
 - When reading from a HID device, check number of bytes returned to detect USB HID timeouts.
 
+- Fix CMSIS-DAP v1 devices on MacOS using non-64-byte report sizes (#681).
+
 ## [0.11.0]
 
 ### Added